// src/ui/display/ColorPicker/ColorPicker.types.ts
export type ColorValue = {
  hue: number;
  saturation: number;
  lightness: number;
  opacity: number;
};

export type HSVaColor = {
  hue: number;
  saturation: number;
  value: number;
  alpha: number;
};

export type HSLAColor = {
  hue: number;
  saturation: number;
  lightness: number;
  alpha: number;
};